Moved test domain objects to src/main/java
