--- ui/tests/src/test/java/org/jboss/forge/addon/ui/enhancer/Gender.java
@@ -1,13 +1,0 @@
-/*
- * Copyright 2013 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Eclipse Public License version 1.0, available at
- * http://www.eclipse.org/legal/epl-v10.html
- */
-
-package org.jboss.forge.addon.ui.enhancer;
-
-public enum Gender
-{
-   MALE, FEMALE;
-}
