+++ ui/tests/src/test/java/org/jboss/forge/addon/ui/enhancer/MainWizard.java
@@ -11,6 +11,7 @@ import javax.inject.Inject;
 import org.jboss.forge.addon.ui.context.UIBuilder;
 import org.jboss.forge.addon.ui.context.UIContext;
 import org.jboss.forge.addon.ui.context.UIValidationContext;
+import org.jboss.forge.addon.ui.domain.Gender;
 import org.jboss.forge.addon.ui.input.SelectComponentEnhancer;
 import org.jboss.forge.addon.ui.input.UIInput;
 import org.jboss.forge.addon.ui.metadata.UICommandMetadata;
