+++ ui/tests/src/test/java/org/jboss/forge/addon/ui/enhancer/SelectComponentEnhancerTest.java
@@ -16,6 +16,7 @@ import javax.inject.Inject;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.forge.addon.ui.domain.Gender;
 import org.jboss.forge.addon.ui.input.InputComponent;
 import org.jboss.forge.addon.ui.input.SelectComponentEnhancer;
 import org.jboss.forge.addon.ui.input.UISelectOne;
