+++ ui/tests/src/test/java/org/jboss/forge/addon/ui/impl/UIInputInjectionTest.java
@@ -16,6 +16,8 @@ import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.forge.addon.convert.Converter;
 import org.jboss.forge.addon.ui.context.UIContext;
+import org.jboss.forge.addon.ui.domain.Career;
+import org.jboss.forge.addon.ui.domain.Gender;
 import org.jboss.forge.addon.ui.facets.HintsFacet;
 import org.jboss.forge.addon.ui.hints.InputType;
 import org.jboss.forge.addon.ui.input.InputComponent;
